extracted useFacetagging from FacetaggingContext to get rid of the HMR-warning, renamed FacetaggingContext to FacetaggingProvider and changed its Folder to make it consistent with other Providers, did some minor refactoring related to these changes
