+++ projects/bp-gallery/src/components/views/picture/PictureView.tsx
@@ -19,7 +19,8 @@ import usePrefetchPictureHook from '../../../hooks/prefetch.hook';
 import usePresentationChannel from '../../../hooks/presentation-channel.hook';
 import { FlatPicture } from '../../../types/additionalFlatTypes';
 import { AuthRole, useAuth } from '../../provider/AuthProvider';
-import { FaceTaggingProvider, FaceTags } from './face-tagging/FaceTaggingContext';
+import { FaceTaggingProvider } from '../../provider/FaceTaggingProvider';
+import { FaceTags } from './face-tagging/FaceTags';
 import { getNextPictureId, getPreviousPictureId } from './helpers/next-prev-picture';
 import { PictureNavigationTarget } from './overlay/PictureNavigationButtons';
 import PictureViewUI from './overlay/PictureViewUI';
